requirements.txt
Flask==3.0.0
Flask-Limiter==3.5.0
gunicorn==21.2.0
google-cloud-datastore==2.19.0
google-cloud-storage==2.14.0
PyJWT==2.8.0
sendgrid==6.11.0
python-dotenv==1.0.0
scikit-learn==1.4.2
numpy==1.26.4
anthropic==0.28.0
